// lib/game-config/generated.ts
export const cards = [
  {
    "name": "R1",
    "rarity": "R",
    "image": "/images/cards/player/R1/card.webp"
  },
  {
    "name": "R2",
    "rarity": "R",
    "image": "/images/cards/player/R2/card.webp"
  },
  {
    "name": "R3",
    "rarity": "R",
    "image": "/images/cards/player/R3/card.webp"
  },
  {
    "name": "R4",
    "rarity": "R",
    "image": "/images/cards/player/R4/card.webp"
  },
  {
    "name": "SR1",
    "rarity": "SR",
    "image": "/images/cards/player/SR1/card.webp"
  },
  {
    "name": "SR2",
    "rarity": "SR",
    "image": "/images/cards/player/SR2/card.webp"
  },
  {
    "name": "SR3",
    "rarity": "SR",
    "image": "/images/cards/player/SR3/card.webp"
  },
  {
    "name": "SR4",
    "rarity": "SR",
    "image": "/images/cards/player/SR4/card.webp"
  },
  {
    "name": "SSR1",
    "rarity": "SSR",
    "image": "/images/cards/player/SSR1/card.webp"
  },
  {
    "name": "SSR2",
    "rarity": "SSR",
    "image": "/images/cards/player/SSR2/card.webp"
  },
  {
    "name": "SSR3",
    "rarity": "SSR",
    "image": "/images/cards/player/SSR3/card.webp"
  },
  {
    "name": "SSR4",
    "rarity": "SSR",
    "image": "/images/cards/player/SSR4/card.webp"
  },
  {
    "name": "UR1",
    "rarity": "UR",
    "image": "/images/cards/player/UR1/card.webp"
  },
  {
    "name": "UR2",
    "rarity": "UR",
    "image": "/images/cards/player/UR2/card.webp"
  },
  {
    "name": "UR3",
    "rarity": "UR",
    "image": "/images/cards/player/UR3/card.webp"
  }
] as const;

export const rarityRates = {
  "R": 60.0,
  "SR": 25.0,
  "SSR": 10.0,
  "UR": 5.0
} as const;